Fix IMAP connection hang forever
I'm using OTRS and it tries to fetch mails from an IMAPS (port 993) account. It uses Net:IMAP:Simple for this task. Sometimes the process hangs forever and it is always right after this warning:
[...cpan-lib/Net/IMAP/Simple.pm line 1197 in sub _seterrstr] warning unknown return string (id=8): 7 OK EXPUNGE completed.\r\n

I've run tcpdump over this communication and no packet is sent either way because both, client and server already sent what was suposed to be sent. So, after this, the only way out is timeout from one side. It happens because the code does not wait command response to be processed.

This do not happen to be all the time, but just some attempts.

Changing this line fix this problem.